Move layer definitions to the layers/ directory
