+++ frontend/src/layers/CumuliDepth.tsx
@@ -2,8 +2,9 @@ import { Forecast, ForecastPoint } from "../data/Forecast";
 import { ColorScale, Color } from "../ColorScale";
 import * as L from 'leaflet';
 import { Renderer } from "../map/CanvasLayer";
+import { colorScaleEl, Layer } from "./Layer";
 
-export const colorScale = new ColorScale([
+const colorScale = new ColorScale([
   [50,   new Color(0xff, 0xff, 0xff, 0)],
   [400,  new Color(0xff, 0xff, 0xff, 0.25)],
   [800,  new Color(0xff, 0xff, 0xff, 0.5)],
@@ -11,9 +12,7 @@ export const colorScale = new ColorScale([
   [3000, new Color(0xff, 0x00, 0x00, 0.5)]
 ]);
 
-
-
-export class CumuliDepth implements Renderer {
+class CumuliDepth implements Renderer {
 
   constructor(readonly forecast: Forecast) {}
 
@@ -34,3 +33,10 @@ export class CumuliDepth implements Renderer {
   }
 
 }
+
+export const cumuliDepthLayer = new Layer(
+  'Convective Clouds',
+  'Convective Clouds Depth',
+  forecast => new CumuliDepth(forecast),
+  colorScaleEl(colorScale, value => `${value} m `)
+);
