+++ frontend/src/layers/ThermalVelocity.tsx
@@ -2,8 +2,9 @@ import { Forecast, ForecastPoint } from "../data/Forecast";
 import * as L from 'leaflet';
 import { ColorScale, Color } from "../ColorScale";
 import { Renderer } from "../map/CanvasLayer";
+import { colorScaleEl, Layer } from "./Layer";
 
-export class ThermalVelocity implements Renderer {
+class ThermalVelocity implements Renderer {
 
   constructor(readonly forecast: Forecast) {}
 
@@ -21,7 +22,7 @@ export class ThermalVelocity implements Renderer {
 
 }
 
-export const thermalVelocityColorScale = new ColorScale([
+const thermalVelocityColorScale = new ColorScale([
   [0.25, new Color(0x33, 0x33, 0x33, 1)],
   [0.50, new Color(0x99, 0x00, 0x99, 1)],
   [0.75, new Color(0xff, 0x00, 0x00, 1)],
@@ -33,3 +34,10 @@ export const thermalVelocityColorScale = new ColorScale([
   [2.50, new Color(0x99, 0xff, 0xff, 1)],
   [3.00, new Color(0xff, 0xff, 0xff, 1)]
 ]);
+
+export const thermalVelocityLayer = new Layer(
+  'Thermal Velocity',
+  'Thermal updraft velocity',
+  forecast => new ThermalVelocity(forecast),
+  colorScaleEl(thermalVelocityColorScale, value => `${value} m/s `)
+);
