+++ frontend/src/layers/Wind.tsx
@@ -2,8 +2,9 @@ import { drawWindArrow } from "../shapes";
 import { Forecast, ForecastPoint } from "../data/Forecast";
 import * as L from 'leaflet';
 import { Renderer } from "../map/CanvasLayer";
+import { Layer, windColor, windScaleEl } from "./Layer";
 
-export class Wind implements Renderer {
+class Wind implements Renderer {
 
   constructor(readonly forecast: Forecast, readonly wind: ((forecast: ForecastPoint) => [number, number])) {}
 
@@ -24,4 +25,30 @@ export class Wind implements Renderer {
 
 }
 
-export const windColor = (opacity: number): string => `rgba(62, 0, 0, ${opacity})`;
+export const surfaceWindLayer = new Layer(
+  'Surface',
+  'Wind force and direction on the ground',
+  forecast => new Wind(forecast, (point) => [point.uSurfaceWind, point.vSurfaceWind]),
+  windScaleEl
+);
+
+export const _300MAGLWindLayer = new Layer(
+  '300 m AGL',
+  'Wind force and direction at 300 m above the ground level',
+  forecast => new Wind(forecast, (forecast) => [forecast.u300MWind, forecast.v300MWind]),
+  windScaleEl
+);
+
+export const boundaryLayerWindLayer = new Layer(
+  'Boundary Layer',
+  'Average wind force and direction in the boundary layer',
+  forecast => new Wind(forecast, (point) => [point.uWind, point.vWind]),
+  windScaleEl
+);
+
+export const boundaryLayerTopWindLayer = new Layer(
+  'Boundary Layer Top',
+  'Wind force and direction at the top of the boundary layer',
+  forecast => new Wind(forecast, (point) => [point.uBLTopWind, point.vBLTopWind]),
+  windScaleEl
+);
